Fix `sdl2_backend=disabled`
* main: add missing condition in the Wayland fallback.
* sdlwindow: move out 'GetX11HostCursor()' into a separate file.
* build: add 'x11cursor.cpp'.
