+++ src/meson.build
@@ -108,6 +108,7 @@ src = [
   'mangoapp.cpp',
   'reshade_effect_manager.cpp',
   'backend.cpp',
+  'x11cursor.cpp',
   'wayland_backend.cpp',
   'InputEmulation.cpp',
 ]
